requirements.txt
certifi>=2020.12.5
cycler>=0.10.0
kiwisolver>=1.3.1
matplotlib>=3.4.0
numpy>=1.20.2
opencv-python>=4.5.1.48
Pillow>=8.1.2
pyparsing>=2.4.7
python-dateutil>=2.8.1
six>=1.15.0
imutils>=0.5.4
python-vlc>=3.0.21203